SMTC integrates with crypto card based on atf-slibce engine;add sdf, TSAPI and etc.
DRBG set reseed interval and reseed_time_interval according to GM/T 0105-2021.

Modify of the integrity verification mechanism to SM2 signature.

App rand supports getting entropy from specified source.

Rename entropy rtc1 to rtcode, rtc2 to rtmem.

Add rtsock entropy source.

Support requesting entropy from specified entropy source.

Support atf-slibce engine.

Add sdf framework and partial APIs.

Add sdf app.

App speed add keygen for SM2 and GenerateKey with SDF.

Add TSAPI, mainly includes functions such as random number,

signature verification, encryption and decryption, etc.

+++ ssl/statem_ntls/ntls_statem_srvr.c
@@ -1584,7 +1584,7 @@ int tls_construct_server_key_exchange_ntls(SSL *s, WPACKET *pkt)
 {
     unsigned char *encodedPoint = NULL;
     size_t encodedlen = 0;
-    int curve_id = 0;
+    uint16_t curve_id = 0;
     const SIGALG_LOOKUP *lu = s->s3.tmp.sigalg;
     unsigned long type;
     EVP_MD_CTX *md_ctx = EVP_MD_CTX_new();
